backend/requirements.txt
# Flask and extensions
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# OpenAI
openai>=1.10.0

# Utilities
python-dotenv==1.0.1
Werkzeug==3.0.1
PyJWT==2.8.0
PyPDF2==3.0.1
python-pptx==0.6.23
requests
gunicorn
psycopg2-binary